backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
psycopg2-binary>=2.9.9

